Board - User 연관관계 추가, 관련 기능 수정 (username 출력 추가)

+++ src/main/java/com/sparta/BoardAPI2/service/BoardService.java
@@ -9,11 +9,14 @@ import com.sparta.BoardAPI2.repository.BoardRepository;
 import com.sparta.BoardAPI2.repository.UserRepository;
 import com.sparta.BoardAPI2.security.UserDetailsImpl;
 import lombok.RequiredArgsConstructor;
+import org.springframework.security.core.Authentication;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 
 import javax.transaction.Transactional;
 import java.util.List;
+import java.util.Optional;
 
 @RequiredArgsConstructor
 @Service
@@ -24,11 +27,11 @@ public class BoardService {
 
     // 글 생성 (user 매핑)
 //    @Transactional
-//    public BoardResponseDto createBoard(User user, BoardRequestDto requestDto) {
-//        Board board = new Board(user, requestDto);
-//        boardRepository.save(board);
-//        return new BoardResponseDto(board);
-//    }
+    public BoardResponseDto createBoard(User user, BoardRequestDto requestDto) {
+        Board board = new Board(user, requestDto);
+        boardRepository.save(board);
+        return new BoardResponseDto(user, board);
+    }
 
 //    public BoardResponseDto createBoard(@AuthenticationPrincipal UserDetailsImpl userDetails, BoardRequestDto requestDto) {
 //        Board board = new Board(userDetails, requestDto);
@@ -36,11 +39,20 @@ public class BoardService {
 //        return new BoardResponseDto(board);
 //    }
 // 글 생성
-public BoardResponseDto createBoard(BoardRequestDto requestDto) {
-    Board board = new Board(requestDto);
-    boardRepository.save(board);
-    return new BoardResponseDto(board);
-}
+//public BoardResponseDto createBoard(BoardRequestDto requestDto) {
+//    Board board = new Board(requestDto);
+//    Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+//    boardRepository.save(board);
+//    return new BoardResponseDto(board);
+//}
+
+//    public BoardResponseDto createBoard(BoardRequestDto requestDto) {
+//        Board board = new Board(requestDto);
+//        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+////        requestDto.setUsername(authentication.getName());
+//        boardRepository.save(board);
+//        return new BoardResponseDto(board);
+//    }
 
     // 모든 글 가져오기
 //    public List<BoardResponseDto> findAllBoard() {
@@ -67,26 +79,39 @@ public BoardResponseDto createBoard(BoardRequestDto requestDto) {
     // 글 하나 가져오기
     public BoardResponseDto findOneBoard(Long id) {
         Board board = boardRepository.findById(id).orElseThrow(
-                () -> new IllegalArgumentException("조회 실패")
+                () -> new IllegalArgumentException("게시글이 없습니다")
         );
         return new BoardResponseDto(board);
     }
 
     // 글 수정
     @Transactional
-    public Long updateBoard(Long id, BoardRequestDto requestDto) {
+    public Long updateBoard(User user, Long id, BoardRequestDto requestDto) {
         Board board = boardRepository.findById(id).orElseThrow(
-                () -> new IllegalArgumentException("해당 아이디가 존재하지 않습니다.")
+                () -> new IllegalArgumentException("해당 게시글이 존재하지 않습니다.")
         );
-        board.update(requestDto);
-        return board.getId();
+        if (user.getUsername().equals(board.getUser().getUsername())) {
+            board.update(requestDto);
+            return board.getId();
+        } else {
+            throw new IllegalArgumentException("작성자만 수정할 수 있습니다");
+        }
     }
 
     // 삭제
     @Transactional
-    public Long deleteBoard(Long id) {
-        boardRepository.deleteById(id);
-        return id;
+    public Long deleteBoard(User user, Long id) {
+        // 어떤 게시판인지 찾기
+        Board board = boardRepository.findById(id).orElseThrow(
+                () -> new IllegalArgumentException("해당 게시글이 존재하지 않습니다.")
+        );
+
+        if (user.getUsername().equals(board.getUser().getUsername())) {
+            boardRepository.deleteById(id);
+            return id;
+        } else {
+            throw new IllegalArgumentException("작성자만 수정할 수 있습니다");
+        }
     }
 
     // 비밀번호 일치 확인
